Fix segfault inside redraw.
Previously, it was possible to navigate to an already-expanded EXPAND
button at the top of the request builder page, which would cause a call
to redraw() before checking whether we needed to scroll again.

This commit protects the call to redraw from segfaulting, by adding a
null check to the current object.