Use heuristics to curses and cdk format markers.
Based on the source code, cdk appears to interpret markers that begin
with `<` unless they are escaped by `\`.

Empirical evidence suggests that characters after `<` are eaten by
either CDK or curses if they are not escaped.

This commit implements both heuristics to try to preserve raw output as
much as possible.